+++ CHANGES.md
@@ -2,6 +2,7 @@
 
 #### 0.2.10
 * Only configure remote repository if contextURL, user, and password are set
+* Fixes missing property exception in case publishing to Maven repository task is not configured yet
 
 #### 0.2.9
 * Adds shortened plugin name 'com.rapidminer.gradle-plugin' to comply with plugins.gradle.org portal standards